requirements.txt
Flask==3.0.3
gunicorn==21.2.0
requests==2.31.0
gspread==6.1.2
google-auth==2.35.0
python-dotenv==1.0.1
